refact(week,scrape): mv Week[Option] to types.week
Prior to this commit, the `Week` and `WeekOption` classes were defined
in the `scrape` module. However, the `scrape` module requires
`selenium`. This means that `selenium` must be installed even if the
scraping functionality is not being used.

In this commit, move the `Week` and `WeekOption` classes to a new
`types.week` module. This makes it possible for code to use the classes
without any requirement for `selenium` to be installed.