fix(legacy/to_legacy): skip empty day menus
Prior to this commit, the `to_legacy()` attempted to process day menus
for which no dishes were selected (i.e., no main dish and no sides).
This lead to errors like:

    AttributeError: 'NoneType' object has no attribute 'name'

since the corresponding menu object was `None`, and thus, did not have
_any_ attributes.

In this commit, modify the `to_legacy()` function to _skip_ day menus
for which no dishes are selected.