fix(scrape): maximize win to avoid splash dialog
Prior to this commit, the Chrome driver was not configured to maximuize
the window before loading the Dinner Daily website. At least as of
2021.11.12, this results in the log in text fields remaining
non-interactable until a notification dialog is closed by clicking
"OK".

In this commit, configure the Chrome driver to maximize the window
before loading the Dinner Daily website. This prevents the notification
dialog from being displayed.