fix(scrape): use ENTER key to submit form
Prior to this commit, the login button was used to submit the user name
and password form. However, at least as of 2021.11.12, the login button
is no longer interactable, resulting in an exception like the following:

    selenium.common.exceptions.ElementNotInteractableException: Message: element not interactable: element has zero size

In this commit, rather than using the login button was used to submit
the user name and password form, instead submit the form by sending the
ENTER key to the password text input. This eliminates the dependence on
the login button all together.