fix(scrape): query shadow selector in Javascript
Prior to this commit, the `execute_script()` Selenium function was used
to directly return the shadow root corresponding to the specified
element. However, at least as of `selenium==3.141.0` and
`python-chromedriver-binary==96.0.4664.35.0`, the `shadowRoot`
Javascript proprerty of an element now returns a dictionary through the
`WebDriver.execute_script()` API. This change in behaviour breaks the
`shadow()` function in the `scrape` module.

In this commit, work around the issue by replacing the `shadow()`
function in the `scrape` module with a new `scrape_query_selector()`
function, which directly executes a `shadowRoot.querySelector(...)` call
through the `WebDriver.execute_script()` API instead. Also update all
existing `shadow()` calls to use the new `scrape_query_selector()`
function.